Separate the auto gui functionality for Qt
